requirements.txt
feedparser
requests
openai
python-dotenv
flask
